Irda: Add Irda System Service for I9505
Add Irda System Service that is compatible with the Galaxy S4.
This is not a perfect solution yet, but it does work and should therefore be a starting point.
I do not know how to make this code exclusive to the S4, maybe someone can help me out with that.

PatchSet 2: Now the Irda service is not started by default. It can be enabled via overlay.

PatchSet 3: Rebase

Change-Id: Id4da55cfbd297fa46c50aa8a6af9cd4d056086cd

Conflicts:
	core/java/android/app/ContextImpl.java
	core/res/res/values/config.xml
	core/res/res/values/symbols.xml
	services/java/com/android/server/SystemServer.java